requirements.txt
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.10.4
attrs==25.4.0
blinker==1.9.0
brotli==1.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
ddgs==9.10.0
distro==1.9.0
discord.py==2.6.3
duckduckgo_search==8.1.1
fake-useragent==2.2.0
Flask==3.1.3
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
lxml==6.0.2
MarkupSafe==3.0.3
newsapi-python==0.2.7
openai==2.21.0
outcome==1.3.0.post0
packaging==26.0
primp==1.0.0
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
PySocks==1.7.1
python-dotenv==1.2.1
python-telegram-bot==22.6
pytz==2025.2
requests==2.32.5
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
tqdm==4.67.3
trio==0.33.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.9.0
Werkzeug==3.1.6
Wikipedia-API==0.9.0
wsproto==1.3.2
